Organize RSA files, remove multi-tx verification
