+++ programs/solana-aa/src/contract/auth/rsa/mod.rs
@@ -1,5 +1,4 @@
 pub mod constants;
-pub mod rsa_multi_tx;
 pub mod rsa_native;
 pub mod rsa_rsa_crate;
 pub mod utils;
