--- programs/solana-aa/src/contract/auth/rsa/rsa_multi_tx.rs
@@ -1,336 +1,0 @@
-use anchor_lang::prelude::*;
-use rsa::pkcs1::DecodeRsaPublicKey;
-use rsa::traits::PublicKeyParts;
-use rsa::RsaPublicKey;
-
-use super::constants::*;
-use super::utils::{ErrorCode, OidcProvider, OidcVerificationData};
-
-/// State for multi-transaction modular exponentiation
-#[account]
-pub struct ModpowState {
-    /// Unique identifier for this modpow operation
-    pub operation_id: [u8; 32],
-    /// Current result of the modpow operation
-    pub current_result: Vec<u8>,
-    /// Remaining exponent bits to process
-    pub remaining_exponent: Vec<u8>,
-    /// Modulus for the operation
-    pub modulus: Vec<u8>,
-    /// Base for the operation
-    pub base: Vec<u8>,
-    /// Current bit position in the exponent
-    pub current_bit_position: u32,
-    /// Total bits in the exponent
-    pub total_bits: u32,
-    /// Whether the operation is complete
-    pub is_complete: bool,
-    /// Original verification data hash for integrity
-    pub verification_data_hash: [u8; 32],
-    /// PDA bump
-    pub bump: u8,
-}
-
-impl ModpowState {
-    pub const DISCRIMINATOR_SIZE: usize = 8;
-    pub const OPERATION_ID_SIZE: usize = 32;
-    pub const VEC_PREFIX_SIZE: usize = 4;
-    pub const U32_SIZE: usize = 4;
-    pub const BOOL_SIZE: usize = 1;
-    pub const HASH_SIZE: usize = 32;
-    pub const BUMP_SIZE: usize = 1;
-
-    /// Calculate the space needed for a ModpowState account
-    pub fn calculate_space(
-        result_size: usize,
-        exponent_size: usize,
-        modulus_size: usize,
-        base_size: usize,
-    ) -> usize {
-        Self::DISCRIMINATOR_SIZE
-            + Self::OPERATION_ID_SIZE
-            + Self::VEC_PREFIX_SIZE + result_size
-            + Self::VEC_PREFIX_SIZE + exponent_size
-            + Self::VEC_PREFIX_SIZE + modulus_size
-            + Self::VEC_PREFIX_SIZE + base_size
-            + Self::U32_SIZE * 2 // current_bit_position + total_bits
-            + Self::BOOL_SIZE
-            + Self::HASH_SIZE
-            + Self::BUMP_SIZE
-    }
-}
-
-/// Initialize a multi-transaction RSA verification
-pub fn init_rsa_verification(
-    ctx: Context<InitRsaVerification>,
-    verification_data: OidcVerificationData,
-) -> Result<()> {
-    verification_data.validate()?;
-
-    // Get public key based on provider and key index
-    let public_key_der = match verification_data.provider {
-        OidcProvider::Google => GOOGLE_RSA_PUBLIC_KEYS[verification_data.key_index as usize],
-    };
-
-    // Parse DER-encoded public key
-    let public_key = RsaPublicKey::from_pkcs1_der(public_key_der)
-        .map_err(|_| error!(ErrorCode::InvalidDerEncoding))?;
-
-    // Extract RSA components
-    let modulus_bytes = public_key.n().to_bytes_be();
-    let exponent_bytes = public_key.e().to_bytes_be();
-
-    // Calculate verification data hash for integrity
-    let verification_data_hash = {
-        use anchor_lang::solana_program::hash::hash;
-        let mut hasher_input = Vec::new();
-        hasher_input.extend_from_slice(&verification_data.signing_input_hash);
-        hasher_input.extend_from_slice(&verification_data.signature);
-        hasher_input.push(verification_data.key_index);
-        let hash_result = hash(&hasher_input);
-        hash_result.to_bytes()
-    };
-
-    // Initialize the modpow state
-    let state = &mut ctx.accounts.modpow_state;
-    state.operation_id = ctx.accounts.operation_id.key().to_bytes();
-    state.current_result = vec![1u8]; // Start with 1
-    state.remaining_exponent = exponent_bytes.clone();
-    state.modulus = modulus_bytes;
-    state.base = verification_data.signature;
-    state.current_bit_position = 0;
-    state.total_bits = (exponent_bytes.len() * 8) as u32;
-    state.is_complete = false;
-    state.verification_data_hash = verification_data_hash;
-    state.bump = ctx.bumps.modpow_state;
-
-    msg!(
-        "Initialized RSA verification with {} bits",
-        state.total_bits
-    );
-    Ok(())
-}
-
-/// Continue the modular exponentiation for a number of iterations
-pub fn continue_rsa_verification(ctx: Context<ContinueRsaVerification>) -> Result<bool> {
-    let state = &mut ctx.accounts.modpow_state;
-
-    if state.is_complete {
-        return Ok(true);
-    }
-
-    // Process only 1 bit per transaction to minimize memory usage
-    if state.current_bit_position >= state.total_bits {
-        state.is_complete = true;
-        msg!("RSA modpow completed after {} total bits", state.total_bits);
-        return Ok(true);
-    }
-
-    let byte_index = (state.current_bit_position / 8) as usize;
-    let bit_index = 7 - (state.current_bit_position % 8);
-
-    if byte_index >= state.remaining_exponent.len() {
-        state.is_complete = true;
-        return Ok(true);
-    }
-
-    let bit = (state.remaining_exponent[byte_index] >> bit_index) & 1;
-
-    // Use a simplified approach that avoids BigUint allocations
-    // This is a proof-of-concept - NOT cryptographically correct
-    let base_clone = state.base.clone();
-    let modulus_clone = state.modulus.clone();
-    let _result_changed = perform_simplified_bit_operation(
-        &mut state.current_result,
-        &base_clone,
-        &modulus_clone,
-        bit,
-    )?;
-
-    state.current_bit_position += 1;
-
-    // Check if we're done
-    if state.current_bit_position >= state.total_bits {
-        state.is_complete = true;
-        msg!("RSA modpow completed after {} total bits", state.total_bits);
-    } else {
-        msg!(
-            "Processed bit {}/{} (bit value: {})",
-            state.current_bit_position,
-            state.total_bits,
-            bit
-        );
-    }
-
-    Ok(state.is_complete)
-}
-
-/// Simplified single bit operation to minimize memory usage
-/// WARNING: This is NOT cryptographically correct - just a proof of concept
-fn perform_simplified_bit_operation(
-    current_result: &mut Vec<u8>,
-    base: &[u8],
-    _modulus: &[u8],
-    bit: u8,
-) -> Result<bool> {
-    // Ensure result has enough space
-    if current_result.len() < 256 {
-        current_result.resize(256, 0);
-    }
-
-    // Simple operations that don't use BigUint
-    // This is NOT correct RSA but shows the memory management works
-    if bit == 1 {
-        // Simple addition operation
-        for i in 0..current_result.len().min(base.len()) {
-            current_result[i] = current_result[i].wrapping_add(base[i]);
-        }
-    }
-
-    // Simple "squaring" operation
-    for i in 0..128 {
-        if i < current_result.len() {
-            current_result[i] = current_result[i].wrapping_mul(2);
-        }
-    }
-
-    Ok(true)
-}
-
-/// Finalize the RSA verification and return the result
-pub fn finalize_rsa_verification(
-    ctx: Context<FinalizeRsaVerification>,
-    verification_data: OidcVerificationData,
-) -> Result<bool> {
-    let state = &ctx.accounts.modpow_state;
-
-    if !state.is_complete {
-        return Err(error!(ErrorCode::ModpowNotComplete));
-    }
-
-    // Verify integrity by checking the verification data hash
-    let verification_data_hash = {
-        use anchor_lang::solana_program::hash::hash;
-        let mut hasher_input = Vec::new();
-        hasher_input.extend_from_slice(&verification_data.signing_input_hash);
-        hasher_input.extend_from_slice(&verification_data.signature);
-        hasher_input.push(verification_data.key_index);
-        let hash_result = hash(&hasher_input);
-        hash_result.to_bytes()
-    };
-
-    if verification_data_hash != state.verification_data_hash {
-        return Err(error!(ErrorCode::VerificationDataMismatch));
-    }
-
-    // For this proof of concept, return a deterministic result based on the final state
-    // In a real implementation, you'd validate PKCS#1 padding here
-    let checksum = state
-        .current_result
-        .iter()
-        .fold(0u8, |acc, &x| acc.wrapping_add(x));
-    let verification_result = checksum > 128; // Simplified check
-
-    Ok(verification_result)
-}
-
-/// Clean up the modpow state account
-pub fn cleanup_rsa_verification(ctx: Context<CleanupRsaVerification>) -> Result<()> {
-    msg!("Cleaned up RSA verification state");
-    Ok(())
-}
-
-#[derive(Accounts)]
-#[instruction(verification_data: OidcVerificationData)]
-pub struct InitRsaVerification<'info> {
-    #[account(mut)]
-    pub payer: Signer<'info>,
-
-    /// CHECK: This is just used as a unique identifier
-    pub operation_id: AccountInfo<'info>,
-
-    #[account(
-        init,
-        payer = payer,
-        space = ModpowState::calculate_space(256, 32, 256, 256), // Conservative estimates for RSA-2048
-        seeds = [
-            b"rsa_modpow",
-            payer.key().as_ref(),
-            operation_id.key().as_ref()
-        ],
-        bump
-    )]
-    pub modpow_state: Account<'info, ModpowState>,
-
-    pub system_program: Program<'info, System>,
-}
-
-#[derive(Accounts)]
-pub struct ContinueRsaVerification<'info> {
-    #[account(mut)]
-    pub payer: Signer<'info>,
-
-    /// CHECK: This is just used as a unique identifier
-    pub operation_id: AccountInfo<'info>,
-
-    #[account(
-        mut,
-        seeds = [
-            b"rsa_modpow",
-            payer.key().as_ref(),
-            operation_id.key().as_ref()
-        ],
-        bump = modpow_state.bump
-    )]
-    pub modpow_state: Account<'info, ModpowState>,
-}
-
-#[derive(Accounts)]
-#[instruction(verification_data: OidcVerificationData)]
-pub struct FinalizeRsaVerification<'info> {
-    #[account(mut)]
-    pub payer: Signer<'info>,
-
-    /// CHECK: This is just used as a unique identifier
-    pub operation_id: AccountInfo<'info>,
-
-    #[account(
-        mut,
-        seeds = [
-            b"rsa_modpow",
-            payer.key().as_ref(),
-            operation_id.key().as_ref()
-        ],
-        bump = modpow_state.bump
-    )]
-    pub modpow_state: Account<'info, ModpowState>,
-}
-
-#[derive(Accounts)]
-pub struct CleanupRsaVerification<'info> {
-    #[account(mut)]
-    pub payer: Signer<'info>,
-
-    /// CHECK: This is just used as a unique identifier
-    pub operation_id: AccountInfo<'info>,
-
-    #[account(
-        mut,
-        seeds = [
-            b"rsa_modpow",
-            payer.key().as_ref(),
-            operation_id.key().as_ref()
-        ],
-        bump = modpow_state.bump,
-        close = payer
-    )]
-    pub modpow_state: Account<'info, ModpowState>,
-}
-
-#[error_code]
-pub enum MultiTxErrorCode {
-    #[msg("Modpow operation not complete")]
-    ModpowNotComplete,
-    #[msg("Verification data mismatch")]
-    VerificationDataMismatch,
-}
